Added Homescreen Class
My new goal is to make the fretboard interactive, so you can press a note on it and hear the note played from the computer. The first step towards this was creating a window and button s for each note in a grid. I did some rudimentary research on the Java Swing class to accomplish these tasks